ecommerce/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
autopep8==1.6.0

certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
cssselect2==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-easy-pdf==0.1.1
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
future==0.18.2
html5lib==1.1
idna==3.3
lxml==4.8.0
oscrypto==1.3.0
Pillow==9.0.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
PyJWT==2.3.0
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
razorpay==1.3.0
reportlab==3.6.9
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
toml==0.10.2
tqdm==4.63.1
twilio==7.6.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.8
webencodings==0.5.1
xhtml2pdf==0.2.7
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0